feat: automatic migrations
