+++ packages/adapter-pocketbase/src/pocketbase.helpers.ts
@@ -1,4 +1,6 @@
-import Pocketbase, { type Record as PBRecord } from "pocketbase"
+import Pocketbase from "pocketbase"
+import { nextAuthCollections } from "./pocketbase.schema"
+import { loginOpts } from "./pocketbase.types"
 
 const PB_RECORD_KEYS = [
   "created",
@@ -11,6 +13,13 @@ const PB_RECORD_KEYS = [
   "export",
   "isNew",
   "load",
+]
+
+const NEXTAUTH_TABLE_NAMES = [
+  "next_auth_user",
+  "next_auth_account",
+  "next_auth_session",
+  "next_auth_verificationToken",
 ]
 
 // ty supabase
@@ -35,48 +44,36 @@ export function format<TAdapterType>(obj: Record<string, any>): TAdapterType {
   return obj as TAdapterType
 }
 
-export async function adminLogin(
-  pb: Pocketbase,
-  options: { username: string; password: string }
-) {
+export async function adminLogin(pb: Pocketbase, options: loginOpts) {
   return await pb.admins.authWithPassword(options.username, options.password)
 }
 
-export type PocketBaseUser = PBRecord & {
-  name: string
-  image: string
-  email: string
-  emailVerified: string
-  code?: number
+export async function initCollections(pb: Pocketbase, options: loginOpts) {
+  try {
+    await adminLogin(pb, options)
+    await pb.collections.import(nextAuthCollections, false)
+  } catch (_) {
+    throw new Error(
+      "could not automatically create the default next_auth collections"
+    )
+  }
 }
 
-export type PocketBaseSession = PBRecord & {
-  expires: string
-  sessionToken: string
-  userId: string
-  code?: number
-}
+export async function checkCollections(pb: Pocketbase, options: loginOpts) {
+  try {
+    await adminLogin(pb, options)
 
-export type PocketBaseAccount = PBRecord & {
-  userId: string
-  type: string
-  provider: string
-  providerAccountId: string
-  refresh_token: string
-  access_token: string
-  expires_at: string
-  token_type: string
-  scope: string
-  id_token: string
-  session_state: string
-  oauth_token_secret: string
-  oauth_token: string
-  code?: number
-}
+    const collectionNames = (
+      await pb.collections.getFullList(200, {
+        sort: "-created",
+      })
+    ).map((collection) => collection.name)
 
-export type PocketBaseVerificationToken = PBRecord & {
-  identifier: string
-  token: string
-  expires: string
-  code?: number
+    // if all of the table names are NOT present in pocketbase, create them.
+    if (!NEXTAUTH_TABLE_NAMES.every((name) => collectionNames.includes(name))) {
+      await initCollections(pb, options)
+    }
+  } catch (e) {
+    throw new Error("error creating collections in pocketbase")
+  }
 }
