+++ packages/adapter-pocketbase/src/pocketbase.schema.ts
@@ -1,0 +1,363 @@
+import type { Collection } from "pocketbase"
+
+const userSchema: Collection = {
+  id: "qhn8ngzgaw4k3x1",
+  name: "next_auth_user",
+  type: "base",
+  system: true,
+  schema: [
+    {
+      id: "jzcv4fft",
+      name: "name",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "1tzuoeth",
+      name: "email",
+      type: "email",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        exceptDomains: null,
+        onlyDomains: null,
+      },
+    },
+    {
+      id: "gtjm3xag",
+      name: "emailVerified",
+      type: "date",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: "",
+        max: "",
+      },
+    },
+    {
+      id: "kwsjxvn9",
+      name: "image",
+      type: "url",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        exceptDomains: null,
+        onlyDomains: null,
+      },
+    },
+  ],
+  listRule: "@request.auth.id != ''",
+  viewRule: "@request.auth.id != ''",
+  createRule: "@request.auth.id != ''",
+  updateRule: "@request.auth.id != ''",
+  deleteRule: "@request.auth.id != ''",
+  options: {},
+}
+
+const accountSchema: Collection = {
+  id: "kmvbgt47p7bndel",
+  name: "next_auth_account",
+  type: "base",
+  system: true,
+  schema: [
+    {
+      id: "m3zu6rgq",
+      name: "userId",
+      type: "relation",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        collectionId: "qhn8ngzgaw4k3x1",
+        cascadeDelete: true,
+        maxSelect: 1,
+        displayFields: null,
+      },
+    },
+    {
+      id: "papslhfw",
+      name: "type",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "uet14quu",
+      name: "provider",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "j5q5swrw",
+      name: "providerAccountId",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "6izkcmmg",
+      name: "refresh_token",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "vkq6dwzy",
+      name: "access_token",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "rxozyv0h",
+      name: "expires_at",
+      type: "number",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+      },
+    },
+    {
+      id: "8s4adqhe",
+      name: "token_type",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "agktlck0",
+      name: "scope",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "u4i6gwbz",
+      name: "id_token",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "l0ijqeyq",
+      name: "session_state",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "hpc1qbl9",
+      name: "oauth_token_secret",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "zsp5yoed",
+      name: "oauth_token",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+  ],
+  listRule: "@request.auth.id != ''",
+  viewRule: "@request.auth.id != ''",
+  createRule: "@request.auth.id != ''",
+  updateRule: "@request.auth.id != ''",
+  deleteRule: "@request.auth.id != ''",
+  options: {},
+}
+
+const sessionSchema: Collection = {
+  id: "0fbhwxgna5jb0k1",
+  name: "next_auth_session",
+  type: "base",
+  system: true,
+  schema: [
+    {
+      id: "bysaoe92",
+      name: "userId",
+      type: "relation",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        collectionId: "qhn8ngzgaw4k3x1",
+        cascadeDelete: true,
+        maxSelect: 1,
+        displayFields: null,
+      },
+    },
+    {
+      id: "8bwckxpy",
+      name: "expires",
+      type: "date",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: "",
+        max: "",
+      },
+    },
+    {
+      id: "fc4ddcvq",
+      name: "sessionToken",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+  ],
+  listRule: "@request.auth.id != ''",
+  viewRule: "@request.auth.id != ''",
+  createRule: "@request.auth.id != ''",
+  updateRule: "@request.auth.id != ''",
+  deleteRule: "@request.auth.id != ''",
+  options: {},
+}
+
+const verificationTokenSchema: Collection = {
+  id: "b1muh5lapcicift",
+  name: "next_auth_verificationToken",
+  type: "base",
+  system: true,
+  schema: [
+    {
+      id: "uvlrbxmy",
+      name: "identifier",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+    {
+      id: "t1halyno",
+      name: "expires",
+      type: "date",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: "",
+        max: "",
+      },
+    },
+    {
+      id: "v0mjzelg",
+      name: "token",
+      type: "text",
+      system: false,
+      required: false,
+      unique: false,
+      options: {
+        min: null,
+        max: null,
+        pattern: "",
+      },
+    },
+  ],
+  listRule: "@request.auth.id != ''",
+  viewRule: "@request.auth.id != ''",
+  createRule: "@request.auth.id != ''",
+  updateRule: "@request.auth.id != ''",
+  deleteRule: "@request.auth.id != ''",
+  options: {},
+}
+
+export const nextAuthCollections = [
+  userSchema,
+  accountSchema,
+  sessionSchema,
+  verificationTokenSchema,
+]
