+++ packages/adapter-pocketbase/tests/index.test.ts
@@ -1,18 +1,18 @@
 import { runBasicTests } from "@next-auth/adapter-test"
-import { PocketBaseAdapter } from "../src"
+import { PocketbaseAdapter } from "../src"
+import { format } from "../src/pocketbase.helpers"
 import type {
   AdapterAccount,
   VerificationToken,
   AdapterUser,
   AdapterSession,
-} from "next-auth/src/adapters"
-import {
-  type PocketBaseAccount,
-  type PocketBaseSession,
-  type PocketBaseUser,
-  type PocketBaseVerificationToken,
-  format,
-} from "../src/pocketbase.helpers"
+} from "@auth/core/src/adapters"
+import type {
+  PocketBaseAccount,
+  PocketBaseSession,
+  PocketBaseUser,
+  PocketBaseVerificationToken,
+} from "../src/pocketbase.types"
 
 import "cross-fetch/polyfill"
 
@@ -20,16 +20,18 @@ import Pocketbase from "pocketbase"
 const pb = new Pocketbase("http://127.0.0.1:8090")
 
 runBasicTests({
-  adapter: PocketBaseAdapter(pb, {
-    username: "test@test.com",
+  adapter: PocketbaseAdapter(pb, {
+    username: "Test@test.com",
     password: "pocketbase1234",
   }),
   db: {
     async session(sessionToken) {
       try {
-        let pb_session = await pb
+        const pb_session = await pb
           .collection("next_auth_session")
           .getFirstListItem<PocketBaseSession>(`sessionToken="${sessionToken}"`)
+
+        if (pb_session.code) throw new Error("could not find session")
 
         return format<AdapterSession>(pb_session)
       } catch (_) {
@@ -38,7 +40,7 @@ runBasicTests({
     },
     async user(id) {
       try {
-        let pb_user = await pb
+        const pb_user = await pb
           .collection("next_auth_user")
           .getOne<PocketBaseUser>(id)
         if (pb_user.code) throw new Error("could not find user")
@@ -50,7 +52,7 @@ runBasicTests({
     },
     async account({ provider, providerAccountId }) {
       try {
-        let pb_account = await pb
+        const pb_account = await pb
           .collection("next_auth_account")
           .getFirstListItem<PocketBaseAccount>(
             `provider="${provider}" && providerAccountId="${providerAccountId}"`
@@ -69,13 +71,14 @@ runBasicTests({
     },
     async verificationToken({ identifier, token }) {
       try {
-        let pb_veriToken = await pb
+        const pb_veriToken = await pb
           .collection("next_auth_verificationToken")
           .getFirstListItem<PocketBaseVerificationToken>(
             `identifier="${identifier}" && token="${token}"`
           )
+
         if (pb_veriToken.code)
-          throw new Error("could not get verificationToken")
+          throw new Error("could not find verificationToken")
 
         // @ts-expect-error
         const { id, ...verificationToken } =
